Cleanup, add PHP file support (#270)
* Cleanup, add PHP file support

* Update changelog

* Update Behat matrix

* Lint fixes

* PHPStan level 6

* PHPStan level 7

* PHPStan level 8

* Fix requests, PHPStan level 9

* Remove unused imports

* PHStan for tests & lint fixes

* Bump to PHP 8.1

* Fix Behat coverage config

* Fix Behat tests

* PHPStan fixes

* Update changelog

* Fix normalize step

* Fix lint step

* Update command

* Remove php 8.4 step for now

* Apply feedback from code review

* Replace deprecated set-output command

---------

+++ inc/Runner.php
@@ -22,7 +22,7 @@ class Runner {
 	 *
 	 * @var \Required\Traduttore\Loader VCS loader.
 	 */
-	protected $loader;
+	protected Loader $loader;
 
 	/**
 	 * Updater instance.
@@ -31,7 +31,7 @@ class Runner {
 	 *
 	 * @var \Required\Traduttore\Updater Translation updater.
 	 */
-	protected $updater;
+	protected Updater $updater;
 
 	/**
 	 * Runner constructor.
@@ -63,6 +63,10 @@ class Runner {
 			if ( ! \WP_Filesystem() ) {
 				return false;
 			}
+		}
+
+		if ( ! $wp_filesystem ) {
+			return false;
 		}
 
 		return $wp_filesystem->rmdir( $this->loader->get_local_path(), true );
